+++ seeker/src/main/java/se452/group9/seeker/model/Recruiter.java
@@ -1,4 +1,4 @@
-package se452.group9.seeker;
+package se452.group9.seeker.model;
 
 
 import javax.persistence.Column;
