+++ seeker/src/main/java/se452/group9/seeker/repo/DummyJobsRepository.java
@@ -1,5 +1,6 @@
-package se452.group9.seeker;
+package se452.group9.seeker.repo;
 
+import se452.group9.seeker.model.DummyJobs;
 import org.springframework.data.repository.CrudRepository;
 
 public interface DummyJobsRepository extends CrudRepository<DummyJobs,Integer> {
